[front/components/vaults] - refactor: streamline modals for editing vault data sources
 - Consolidate separate folder and website modals into a single component `EditVaultStaticDatasourcesViews`
 - Remove `VaultFolderOrWebsiteModal` component to reduce redundancy and simplify the data source editing logic
 - Introduce new prop `isOpen` along with a handler `setOpen` to manage modal visibility within the unified component
 - Extend `EditVaultStaticDatasourcesViewsProps` interface to include types and handlers for the improved component structure
 - Use conditional rendering based on `category` within `EditVaultStaticDatasourcesViews` to support both folder and website view modifications
